Code added for Auto PST Generator
This Code will automatically Identifies the Sql from given package or
file based on pattern.
Once Pattern is recorgnised it will create the separate file for the
same and run against the database to get the explain plan and other
stats.
Also it will upload all the file in excel and attach the generated
files.